Check Restore error log process

+++ includes/class-instawp-rest-api.php
@@ -424,18 +424,18 @@ class InstaWP_Backup_Api
                $url = $api_doamin . INSTAWP_API_URL . '/connects/' . $id . '/restore_status';
 
                // restore preogress precetage
-               $restore_progress_option = get_option('instawp_restore_progress_percents', "0");
+               $restore_progress_option = (int)get_option('instawp_restore_progress_percents', "0");
                error_log('Restore Status percentage is : '. $restore_progress_option);
 
                $body = array(
-				   "task_id"         => $task_id,
-				   // "type"     => 'restore',
-				   "progress"        => $restore_progress_option,
-				   "message"         => $response['message'],
-				   "connect_id"      => $id,
-				   "completed"       => true,
-				   "status"          => true,
-				   "destination_url" => get_site_url(),
+                  "task_id"         => $task_id,
+                  // "type"     => 'restore',
+                  "progress"        => $restore_progress_option,
+                  "message"         => $response['message'],
+                  "connect_id"      => $id,
+                  "completed"       => true,
+                  "status"          => true,
+                  "destination_url" => get_site_url(),
                );
                $body_json     = json_encode($body);
